requirements.txt
# Chess Analysis Dependencies
# Install with: pip install -r requirements.txt

# Core chess libraries
chess>=1.10.0
chess.com>=3.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
ipywidgets>=8.0.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
nbconvert>=7.0.0
papermill>=2.4.0

# Templating (for HTML reports)
Jinja2>=3.1.0

# Machine learning (for Maia2)
torch>=2.0.0
einops>=0.6.0

# Network analysis (for fragility)
networkx>=3.0

# HTTP requests
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.64.0
pyzstd>=0.15.0

# Optional: scipy for statistical analysis
scipy>=1.10.0
